ps3 working with bp32
