+++ OASMan_ESP32/platformio.ini
@@ -38,7 +38,7 @@ lib_deps =
     ; jvpernis/PS3 Controller Host@^1.1.0
     ; ;
 
-monitor_port = COM6
-upload_port = COM6
+;monitor_port = COM6
+;upload_port = COM6
 
 extra_scripts = post:post_build_script.py
