+++ OASMan_ESP32/src/bluetooth/ble.h
@@ -1,30 +1,29 @@
-#ifndef ble_h
-#define ble_h
+#ifndef ble2_h
+#define ble2_h
 
-//#define OASMAN_BLE
-
-// Tutorial: https://www.youtube.com/watch?v=s3yoZa6kzus&ab_channel=MoThunderz
-// https://github.com/mo-thunderz/Esp32BlePart2/tree/main
-
-/*
-  Based on Neil Kolban example for IDF: https://github.com/nkolban/esp32-snippets/blob/master/cpp_utils/tests/BLE%20Tests/SampleNotify.cpp
-  Ported to Arduino ESP32 by Evandro Copercini
-  updated by chegewara and MoThunderz
-*/
 #include <Arduino.h>
-#include <BLEDevice.h>
-#include <BLEServer.h>
-#include <BLEUtils.h>
-#include <BLE2902.h>
+#include "btstack.h"
+#include "btstack_run_loop.h"
+#include "ble/att_db.h"
+#include "btstack_util.h"
+#include "att_db_util.h"
+#include <string.h>
+
+#include "hci.h"
+#include "gap.h"
+#include "l2cap.h"
+#include "sm.h"
+#include "att_server.h"
+
+#include "btstack_defines.h"
+#include "ble/att_db_util.h"
 
 #include "airSuspensionUtil.h"
 #include <BTOas.h>
 #include "components/manifold.h"
 #include "tasks/tasks.h"
 
-#ifdef OASMAN_BLE
 void ble_setup();
 void ble_loop();
-#endif
 
 #endif
